requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
python-dotenv
langchain
langchain-community
faiss-cpu
sentence-transformers
tavily-python
pypdf